Add files via upload
Welcome to my beginner-friendly portfolio! Here, I'm sharing my early exploration into the world of Web Development and Machine Learning. Take a peek at simple projects reflecting my passion for learning Frontend and exploring new frameworks. With a touch of HTML, CSS, and a sprinkle of JavaScript, this portfolio is a snapshot of my beginner's journey